Swapswap Ver.1.1.1 배포

+++ src/main/java/piglin/swapswap/domain/deal/dto/request/DealUpdateRequestDto.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 public record DealUpdateRequestDto (
         @PositiveOrZero
-        int extraFee,
+        Long extraFee,
         List<Long> postIdList
 ) {
 }
